Update CurrencyConverter.html
Added integration with Lambda that takes results and displays in text box provided.

Additionally, the entries in the box are restricted to numbers and are tested for valid amounts (greater than zero). 
Firefox has issues where type="number" isn't properly reflected but it does work in Chrome / Edge (note: e counts as a number, apparently).